Refactor mock data architecture to centralized factory pattern
Migrate legacy mock data files to a unified lib/mock directory structure. Implement factory functions for Bounties, Projects, Leaderboard, and Wallet entities to improve testability. Centralize exports and resolve associated TypeScript type conflicts in consuming components.
