+++ app/projects/page.tsx
@@ -1,10 +1,9 @@
-import { getAllProjects, getAllProjectTags } from "@/lib/mock-project"
-import { ProjectsDiscovery } from "@/components/projects/projects-discovery"
+import { getAllProjects, getAllProjectTags } from "@/lib/mock";
+import { ProjectsDiscovery } from "@/components/projects/projects-discovery";
 
 export default async function ProjectsPage() {
-  const projects = getAllProjects()
-  const allTags = getAllProjectTags(projects)
+  const projects = getAllProjects();
+  const allTags = getAllProjectTags(projects);
 
-  return <ProjectsDiscovery projects={projects} allTags={allTags} />
+  return <ProjectsDiscovery projects={projects} allTags={allTags} />;
 }
-
