+++ lib/mock/bounties.ts
@@ -1,4 +1,36 @@
+import { mockDiscoverProjects } from "./projects";
 import { Bounty } from "@/types/bounty";
+
+export function makeMockBounty(overrides: Partial<Bounty> = {}): Bounty {
+  return {
+    id: "mock-bounty-" + Math.random().toString(36).substr(2, 9),
+    type: "FIXED_PRICE",
+    title: "Mock Bounty",
+    organizationId: "org-boundless",
+    organization: {
+      id: "org-boundless",
+      name: "Boundless",
+      logo: "/logo.svg",
+      slug: "boundless",
+    },
+    projectId: "proj-bounties",
+    project: {
+      id: "proj-bounties",
+      title: "Bounties Platform",
+      description: null,
+    },
+    githubIssueUrl: "https://github.com/boundlessfi/bounties/issues/1",
+    githubIssueNumber: 1,
+    description: "Mock description",
+    rewardAmount: 500,
+    rewardCurrency: "USDC",
+    status: "OPEN",
+    createdBy: "user-admin",
+    createdAt: new Date().toISOString(),
+    updatedAt: new Date().toISOString(),
+    ...overrides,
+  } as Bounty;
+}
 
 export const mockBounties: Bounty[] = [
   {
@@ -275,3 +307,294 @@ export function getBountyById(id: string): Bounty | undefined {
 export function getAllBounties(): Bounty[] {
   return mockBounties;
 }
+
+export const mockDiscoverBounties: Bounty[] = [
+  {
+    id: "bounty-1",
+    title: "Implement Multi-Signature Wallet Feature",
+    description:
+      "Add multi-signature functionality to existing Stellar wallet with customizable approval thresholds.",
+    type: "FIXED_PRICE",
+    rewardAmount: 5000,
+    rewardCurrency: "USDC",
+    status: "OPEN",
+    organizationId: "org-wallet",
+    organization: {
+      id: "org-wallet",
+      name: "Wallet Project",
+      logo: "/logos/org-wallet.png",
+      slug: "org-wallet",
+    },
+    projectId: mockDiscoverProjects[0].id,
+    project: mockDiscoverProjects[0],
+    githubIssueUrl: "https://github.com/wallet-project/issues/1",
+    githubIssueNumber: null,
+    createdBy: "wallet_project",
+    createdAt: "2024-01-20T00:00:00Z",
+    updatedAt: "2024-01-20T00:00:00Z",
+  },
+  {
+    id: "bounty-2",
+    title: "Design Landing Page for DeFi Protocol",
+    description:
+      "Create modern, responsive landing page design with dark mode support and animated elements.",
+    type: "COMPETITION",
+    rewardAmount: 2000,
+    rewardCurrency: "USDC",
+    status: "OPEN",
+    organizationId: "org-defi",
+    organization: {
+      id: "org-defi",
+      name: "DeFi Startup",
+      logo: "/logos/org-defi.png",
+      slug: "org-defi",
+    },
+    projectId: mockDiscoverProjects[1].id,
+    project: mockDiscoverProjects[1],
+    githubIssueUrl: "https://github.com/defi-startup/issues/1",
+    githubIssueNumber: null,
+    createdBy: "defi_startup",
+    createdAt: "2024-01-19T00:00:00Z",
+    updatedAt: "2024-01-21T00:00:00Z",
+  },
+  {
+    id: "bounty-3",
+    title: "Fix Security Vulnerability in Smart Contract",
+    description:
+      "Identify and fix critical security vulnerability in liquidity pool smart contract.",
+    type: "FIXED_PRICE",
+    rewardAmount: 8000,
+    rewardCurrency: "USDC",
+    status: "IN_PROGRESS",
+    organizationId: "org-security",
+    organization: {
+      id: "org-security",
+      name: "Security Team",
+      logo: "/logos/org-security.png",
+      slug: "org-security",
+    },
+    projectId: mockDiscoverProjects[2].id,
+    project: mockDiscoverProjects[2],
+    githubIssueUrl: "https://github.com/security-team/issues/1",
+    githubIssueNumber: null,
+    createdBy: "security_team",
+    createdAt: "2024-01-18T00:00:00Z",
+    updatedAt: "2024-01-22T00:00:00Z",
+  },
+  {
+    id: "bounty-4",
+    title: "Build Mobile App UI Components",
+    description:
+      "Create reusable React Native components for Stellar wallet mobile app.",
+    type: "MILESTONE_BASED",
+    rewardAmount: 3000,
+    rewardCurrency: "USDC",
+    status: "OPEN",
+    organizationId: "org-mobile",
+    organization: {
+      id: "org-mobile",
+      name: "Mobile Team",
+      logo: "/logos/org-mobile.png",
+      slug: "org-mobile",
+    },
+    projectId: mockDiscoverProjects[3].id,
+    project: mockDiscoverProjects[3],
+    githubIssueUrl: "https://github.com/mobile-team/issues/1",
+    githubIssueNumber: null,
+    createdBy: "mobile_team",
+    createdAt: "2024-01-17T00:00:00Z",
+    updatedAt: "2024-01-17T00:00:00Z",
+  },
+  {
+    id: "bounty-5",
+    title: "Write Integration Tests for DEX",
+    description:
+      "Develop comprehensive integration test suite for decentralized exchange smart contracts.",
+    type: "FIXED_PRICE",
+    rewardAmount: 2500,
+    rewardCurrency: "USDC",
+    status: "OPEN",
+    organizationId: "org-dex",
+    organization: {
+      id: "org-dex",
+      name: "DEX Protocol",
+      logo: "/logos/org-dex.png",
+      slug: "org-dex",
+    },
+    projectId: mockDiscoverProjects[4].id,
+    project: mockDiscoverProjects[4],
+    githubIssueUrl: "https://github.com/dex-protocol/issues/1",
+    githubIssueNumber: null,
+    createdBy: "dex_protocol",
+    createdAt: "2024-01-16T00:00:00Z",
+    updatedAt: "2024-01-20T00:00:00Z",
+  },
+  {
+    id: "bounty-6",
+    title: "Optimize Gas Fees for NFT Minting",
+    description:
+      "Reduce transaction costs for NFT minting operations by optimizing smart contract code.",
+    type: "FIXED_PRICE",
+    rewardAmount: 4000,
+    rewardCurrency: "USDC",
+    status: "COMPLETED",
+    organizationId: "org-nft",
+    organization: {
+      id: "org-nft",
+      name: "NFT Marketplace",
+      logo: "/logos/org-nft.png",
+      slug: "org-nft",
+    },
+    projectId: mockDiscoverProjects[5].id,
+    project: mockDiscoverProjects[5],
+    githubIssueUrl: "https://github.com/nft-marketplace/issues/1",
+    githubIssueNumber: null,
+    createdBy: "nft_marketplace",
+    createdAt: "2024-01-10T00:00:00Z",
+    updatedAt: "2024-01-15T00:00:00Z",
+  },
+  {
+    id: "bounty-7",
+    title: "Create Tutorial Videos for Beginners",
+    description:
+      "Produce 5 tutorial videos explaining Stellar development basics for newcomers.",
+    type: "MILESTONE_BASED",
+    rewardAmount: 1500,
+    rewardCurrency: "USDC",
+    status: "OPEN",
+    organizationId: "org-edu",
+    organization: {
+      id: "org-edu",
+      name: "Education DAO",
+      logo: "/logos/org-edu.png",
+      slug: "org-edu",
+    },
+    projectId: mockDiscoverProjects[6].id,
+    project: mockDiscoverProjects[6],
+    githubIssueUrl: "https://github.com/education-dao/issues/1",
+    githubIssueNumber: null,
+    createdBy: "education_dao",
+    createdAt: "2024-01-15T00:00:00Z",
+    updatedAt: "2024-01-18T00:00:00Z",
+  },
+  {
+    id: "bounty-8",
+    title: "Implement Real-Time Price Oracle",
+    description:
+      "Build reliable price oracle service for DeFi protocols with multiple data sources.",
+    type: "FIXED_PRICE",
+    rewardAmount: 6000,
+    rewardCurrency: "USDC",
+    status: "OPEN",
+    organizationId: "org-oracle",
+    organization: {
+      id: "org-oracle",
+      name: "Oracle Network",
+      logo: "/logos/org-oracle.png",
+      slug: "org-oracle",
+    },
+    projectId: mockDiscoverProjects[7].id,
+    project: mockDiscoverProjects[7],
+    githubIssueUrl: "https://github.com/oracle-network/issues/1",
+    githubIssueNumber: null,
+    createdBy: "oracle_network",
+    createdAt: "2024-01-14T00:00:00Z",
+    updatedAt: "2024-01-21T00:00:00Z",
+  },
+  {
+    id: "bounty-9",
+    title: "Add Dark Mode to Dashboard",
+    description:
+      "Implement dark mode theme with smooth transitions for analytics dashboard.",
+    type: "FIXED_PRICE",
+    rewardAmount: 1000,
+    rewardCurrency: "USDC",
+    status: "IN_PROGRESS",
+    organizationId: "org-analytics",
+    organization: {
+      id: "org-analytics",
+      name: "Analytics Platform",
+      logo: "/logos/org-analytics.png",
+      slug: "org-analytics",
+    },
+    projectId: mockDiscoverProjects[8].id,
+    project: mockDiscoverProjects[8],
+    githubIssueUrl: "https://github.com/analytics-platform/issues/1",
+    githubIssueNumber: null,
+    createdBy: "analytics_platform",
+    createdAt: "2024-01-13T00:00:00Z",
+    updatedAt: "2024-01-19T00:00:00Z",
+  },
+  {
+    id: "bounty-10",
+    title: "Audit Staking Contract",
+    description:
+      "Perform comprehensive security audit of staking smart contract with detailed report.",
+    type: "FIXED_PRICE",
+    rewardAmount: 7000,
+    rewardCurrency: "USDC",
+    status: "OPEN",
+    organizationId: "org-staking",
+    organization: {
+      id: "org-staking",
+      name: "Staking Protocol",
+      logo: "/logos/org-staking.png",
+      slug: "org-staking",
+    },
+    projectId: mockDiscoverProjects[9].id,
+    project: mockDiscoverProjects[9],
+    githubIssueUrl: "https://github.com/staking-protocol/issues/1",
+    githubIssueNumber: null,
+    createdBy: "staking_protocol",
+    createdAt: "2024-01-12T00:00:00Z",
+    updatedAt: "2024-01-22T00:00:00Z",
+  },
+  {
+    id: "bounty-11",
+    title: "Build API Documentation Site",
+    description:
+      "Create interactive API documentation website with code examples and playground.",
+    type: "COMPETITION",
+    rewardAmount: 2800,
+    rewardCurrency: "USDC",
+    status: "OPEN",
+    organizationId: "org-api",
+    organization: {
+      id: "org-api",
+      name: "API Team",
+      logo: "/logos/org-api.png",
+      slug: "org-api",
+    },
+    projectId: mockDiscoverProjects[0].id,
+    project: mockDiscoverProjects[0],
+    githubIssueUrl: "https://github.com/api-team/issues/1",
+    githubIssueNumber: null,
+    createdBy: "api_team",
+    createdAt: "2024-01-11T00:00:00Z",
+    updatedAt: "2024-01-16T00:00:00Z",
+  },
+  {
+    id: "bounty-12",
+    title: "Integrate Wallet Connect",
+    description:
+      "Add Wallet Connect support to DApp for seamless mobile wallet integration.",
+    type: "FIXED_PRICE",
+    rewardAmount: 3500,
+    rewardCurrency: "USDC",
+    status: "OPEN",
+    organizationId: "org-dapp",
+    organization: {
+      id: "org-dapp",
+      name: "DApp Builders",
+      logo: "/logos/org-dapp.png",
+      slug: "org-dapp",
+    },
+    projectId: mockDiscoverProjects[1].id,
+    project: mockDiscoverProjects[1],
+    githubIssueUrl: "https://github.com/dapp-builders/issues/1",
+    githubIssueNumber: null,
+    createdBy: "dapp_builders",
+    createdAt: "2024-01-09T00:00:00Z",
+    updatedAt: "2024-01-20T00:00:00Z",
+  },
+];
