+++ lib/mock/index.ts
@@ -1,0 +1,5 @@
+export * from "./bounties";
+export * from "./projects";
+export * from "./leaderboard";
+export * from "./wallet";
+export * from "./model4";
