+++ package.json
@@ -86,6 +86,7 @@
     "@graphql-codegen/typescript-operations": "^5.0.8",
     "@graphql-codegen/typescript-react-query": "^7.0.0",
     "@graphql-typed-document-node/core": "^3.2.0",
+    "@jest/globals": "^30.4.1",
     "@next/swc-wasm-nodejs": "^16.1.6",
     "@playwright/test": "^1.59.1",
     "@tailwindcss/postcss": "^4",
